requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorlog==6.9.0
Faker==30.8.2
idna==3.10
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
load-dotenv==0.1.0
packaging==24.1
pillow==11.0.0
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.1
setuptools==75.3.0
six==1.16.0
style==1.1.0
svgwrite==1.4.3
typing_extensions==4.12.2
update==0.0.1
urllib3==2.2.3
